+++ src/host/mod.rs
@@ -23,11 +23,14 @@ pub(crate) mod emscripten;
     feature = "jack"
 ))]
 pub(crate) mod jack;
-pub(crate) mod null;
 #[cfg(windows)]
 pub(crate) mod wasapi;
 #[cfg(all(target_arch = "wasm32", feature = "wasm-bindgen"))]
 pub(crate) mod webaudio;
+
+#[cfg(feature = "custom")]
+pub(crate) mod custom;
+pub(crate) mod null;
 
 /// Compile-time assertion that a type implements Send.
 /// Use this macro in each host module to ensure Stream is Send.