+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+- Add support for custom `Host`s, `Device`s, and `Stream`s
 - Add `Sample::bits_per_sample` method.
 - Update `audio_thread_priority` to 0.34.
 - AAudio: Configure buffer to ensure consistent callback buffer sizes.